Merged from Terence side - Updates to Logic System (Not fully implemented), StatsComponent

+++ RogueEngine/Source/WindowHelper.h
@@ -2,8 +2,8 @@
 
 namespace Rogue
 {
-	float GetWindowWidth(HWND hWnd);
-	float GetWindowHeight(HWND hWnd);
+	int GetWindowWidth(HWND hWnd);
+	int GetWindowHeight(HWND hWnd);
 
 	LRESULT CALLBACK WndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 	BOOL WINAPI CtrlHandler(DWORD fdwCtrlType);
